v0.18.0 gitian sigs from midnightmagic
NOTE: this build is significantly later than everyone elses; the -linux sigs
don't match, and on cursory examination this appears to be because of a
version bump in bionic.  The *unsigned* OSX build also doesn't appear to
match up with the other signers; however, the *signed* one does.  Go figure.